Made abstract account which is extended by specific accounts
Only implementation differences between accounts is the type they return in the method "GetAccountType()"

- Plus some missing .jar files which didn't let me run the server correctly ;)
