+++ apache-tomcat-9.0.73/logs/localhost_access_log.2023-05-02.txt
@@ -1,0 +1,28 @@
+127.0.0.1 - - [02/May/2023:14:58:03 +0100] "GET / HTTP/1.1" 200 11428
+127.0.0.1 - - [02/May/2023:14:58:03 +0100] "GET /tomcat.css HTTP/1.1" 200 5895
+127.0.0.1 - - [02/May/2023:14:58:11 +0100] "GET /myApp/manageusers HTTP/1.1" 200 475
+127.0.0.1 - - [02/May/2023:14:58:14 +0100] "GET /myApp/manageusers/login HTTP/1.1" 200 397
+127.0.0.1 - - [02/May/2023:14:58:16 +0100] "GET /myApp/manageusers HTTP/1.1" 200 475
+127.0.0.1 - - [02/May/2023:14:58:17 +0100] "GET /myApp/manageusers/register HTTP/1.1" 200 555
+127.0.0.1 - - [02/May/2023:14:58:18 +0100] "GET /myApp/manageusers HTTP/1.1" 200 475
+127.0.0.1 - - [02/May/2023:14:58:19 +0100] "GET /myApp/manageusers/delete HTTP/1.1" 200 261
+127.0.0.1 - - [02/May/2023:14:58:22 +0100] "GET /myApp/manageusers/changepassword HTTP/1.1" 200 649
+127.0.0.1 - - [02/May/2023:14:58:24 +0100] "GET /myApp/manageusers/logout HTTP/1.1" 404 783
+127.0.0.1 - - [02/May/2023:14:58:43 +0100] "GET /myApp/manageusers/login HTTP/1.1" 200 397
+127.0.0.1 - - [02/May/2023:14:59:09 +0100] "GET /myApp/manageusers HTTP/1.1" 200 475
+127.0.0.1 - - [02/May/2023:14:59:11 +0100] "GET /myApp/manageusers/register HTTP/1.1" 200 555
+127.0.0.1 - - [02/May/2023:14:59:16 +0100] "POST /myApp/manageusers/register HTTP/1.1" 200 611
+127.0.0.1 - - [02/May/2023:14:59:36 +0100] "GET / HTTP/1.1" 200 11428
+127.0.0.1 - - [02/May/2023:14:59:36 +0100] "GET /tomcat.css HTTP/1.1" 200 5895
+127.0.0.1 - - [02/May/2023:14:59:36 +0100] "GET /tomcat.svg HTTP/1.1" 200 68761
+127.0.0.1 - - [02/May/2023:14:59:36 +0100] "GET /bg-nav.png HTTP/1.1" 200 1401
+127.0.0.1 - - [02/May/2023:14:59:36 +0100] "GET /asf-logo-wide.svg HTTP/1.1" 200 27530
+127.0.0.1 - - [02/May/2023:14:59:36 +0100] "GET /bg-button.png HTTP/1.1" 200 713
+127.0.0.1 - - [02/May/2023:14:59:36 +0100] "GET /bg-upper.png HTTP/1.1" 200 3103
+127.0.0.1 - - [02/May/2023:14:59:36 +0100] "GET /bg-middle.png HTTP/1.1" 200 1918
+127.0.0.1 - - [02/May/2023:15:00:23 +0100] "POST /myApp/manageusers/register HTTP/1.1" 200 590
+127.0.0.1 - - [02/May/2023:15:00:25 +0100] "GET /myApp/manageusers HTTP/1.1" 200 475
+127.0.0.1 - - [02/May/2023:15:00:27 +0100] "GET /myApp/manageusers/delete HTTP/1.1" 200 261
+127.0.0.1 - - [02/May/2023:15:00:29 +0100] "GET /myApp/manageusers/register HTTP/1.1" 200 555
+127.0.0.1 - - [02/May/2023:15:00:30 +0100] "GET /myApp/manageusers/register HTTP/1.1" 200 555
+127.0.0.1 - - [02/May/2023:15:00:36 +0100] "POST /myApp/manageusers/register HTTP/1.1" 200 610
