+++ apache-tomcat-9.0.73/webapps/myApp/WEB-INF/data/accounts.txt
@@ -1,3 +1,1 @@
-root,/wIuoZbKYXtTvUkobM+ewg==,root,false,false
-2,rLOOOz65/YFBv1qHVezoTg==,user,true,false
-1,/wIuoZbKYXtTvUkobM+ewg==,user,true,false
+root,PnJtSaIAgf7chh669f933w==,root,false,false
