+++ apache-tomcat-9.0.73/webapps/myApp/src/accounts/RootAccount.java
@@ -1,85 +1,27 @@
 package accounts;
 
-public class RootAccount implements Account, AccountWrite{
-    private String name;
-    private String hash;
-    private String type;
-    private boolean isLoggedIn;
-    private boolean isLocked;
+public class RootAccount extends AccountClass{
 
+    /**
+     * Constructor used for creating accounts (if we ever decide to let the client create more than one root)
+     * @param name Name of the new user account
+     * @param password Password of the new user account
+     */
     public RootAccount(String name, String password) {
-        this.name = name;
-        this.hash = GetHashFromPwd(password);
-        this.type = "user";
-        this.isLoggedIn = false;
-        this.isLocked = false;
-    }
-
-    public RootAccount(String[] parameters) {
-        this.name = parameters[0];
-        this.hash = parameters[1];
-        this.type = parameters[2];
-        this.isLoggedIn = parameters[3].equals("true");
-        this.isLocked = parameters[4].equals("true");
+        super(name, password);
     }
 
     /**
-     * Converts the given string which represents a password to a unconvertible hash that is then stored in the account
-     * @param password String that represents the account's password
-     * @return The hash of the given string
+     * Constructor used for getting the account parameters from the database
+     * @param parameters Data related to the user account
      */
-    private String GetHashFromPwd(String password) {
-        return null;
-    }
-
-    @Override
-    public String GetAccountName() {
-        return this.name;
-    }
-
-    @Override
-    public String GetAccountHash() {
-        return this.hash;
+    public RootAccount(String[] parameters) {
+        super(parameters);
     }
 
     @Override
     public String GetAccountType() {
-        return this.type;
+        return "root";
     }
 
-    @Override
-    public boolean isLoggedIn() {
-        return this.isLoggedIn;
-    }
-
-    @Override
-    public boolean isLocked() {
-        return this.isLocked;
-    }
-
-    @Override
-    public void SetIsLoggedIn(boolean isLoggedIn) {
-        this.isLoggedIn = isLoggedIn;
-    }
-
-    @Override
-    public void SetIsLocked(boolean isLocked) {
-        this.isLocked = isLocked;
-    }
-
-    @Override
-    public void SetAccountName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public void SetAccountHash(String pwd) {
-        this.hash = GetHashFromPwd(pwd);
-    }
-
-    @Override
-    public boolean EqualHash(String pwd) {
-        String hash = GetHashFromPwd(pwd);
-        return this.hash.equals(hash);
-    }
 }
